requirements.txt
python-dotenv
openai
pandas
numpy
faiss-cpu
sounddevice
soundfile